backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
python-dotenv
PyJWT
python-multipart
PyMySQL
faiss-cpu
langchain
langchain-text-splitters
langchain-community
langchain-huggingface
sentence-transformers
pypdf
pandas
python-docx
python-pptx
beautifulsoup4
requests
pillow
pytesseract
groq
openai
ddgs
resend
google-cloud-storage
google-auth